Don't pass event name to event handler
Event name doesn't matter for the final handler. Should never have a
handler for multiple types of events (doing too much).

Plus most of our workers already accept a single object as a param, so I
don't need to do anything to change them. This will just work as is.